Improve test output and fix warnings
This commit introduces two main improvements to the testing setup.

First, the test output is now less verbose by default. This is achieved by configuring `ExUnit` to use the `DotTestReporter` in the `test.exs` configuration file. This makes it easier to see the overall test results without being overwhelmed by a large amount of text.

Second, a number of compiler warnings in the `character_generator_test.exs` file have been fixed. This includes removing unused variables and correcting function signatures. These changes improve the quality of the test suite and make it easier to maintain.